Create Xamarin
In this code, the Animal class has a MakeSound() virtual method, which is then overridden in the Dog and Cat classes. We then create instances of Dog and Cat, but type them as Animal, and call the MakeSound() method, which causes the concrete class's version of the method to be executed.